Hide review UI on resume in Camera2.
PhotoModule is used for capturing intents only on legacy hardware,
and it would show broken UI when paused, then resumed during image
review (confirmation UI).

This CL fixes it by hiding the review UI on resume. This is not ideal,
but it's the same behavior as on non-legacy hardware, and the fix
was straightforward.

Test: Confirmed manually that minimizing and restoring on ARC++
      during image review discards the taken picture.
Bug: 33701231

Change-Id: Ie7f67145fab21c8c1952832bab6cd26591d2bda0
(cherry picked from commit 47dc2d1ed16baac83c6ef14c06c7564422a5967c)