Add refresh all pages of UserGridView if the data is changed.
Sometimes the UserGridView is not refresh completely even the data is changed.
Currently it only called the notifyDataSetChanged() in PagerAdapter. But it is
not engough. It also need override the getItemPosition() to destory all items
and renew them while it is calling the notifyDataSetChanged() to refresh.

Test:
1. Click the overview button and tap on Owner.
2. Add new user by tapping on "Add user" icon.
3. Click the overview button and tap the "Owner" icon to switch user mode
from New user to Owner.
4. Click the overview button and check whether the "Owner", "New user",
"Add guest" and "Add user" icons on quicksetting container.

Jira : None
Test : None