[Email] Fix the loading drawable crash when the display size is small
The best resource should be drawable-v21 when the display size is small.
But the file "./drawable-v21/ic_next_24dp.xml" is a vectorgraph and it
is not matched which required bitmap in file "./drawable/ic_next.xml"

Rename it and the vectorgraph can be used directly when the display
size is small.

Test:
1. Goto Settings->Display->Display size, set to "small".
Or using command: adb shell wm density 136
2. Launch email by command:
adb shell am start com.android.email/.activity.setup.AccountSetupFinal

Test: None
Jira: None