Add user checking when resume playing back
There is no user checking when restoring audio focus to app. If end user
switched to a new user2 after launched 2 media apps and requested audio
focus in both apps in user1. Then try to request and abandon audio focus
in new user2. The media app in user1 will get audio focus. If the app
does not handle the focus properly then we got the issue.

This solution is to add user checking when trying to resume the play.

Test steps:
    1.Owner:launch Local media player>>play a song
    2.Owner:Launch Radio
    3.Switch to New User >> launch Local media player >> play a song
      >> pause the song >> observe

The background music should not start playing.

Then go to CarSettings adjust the volume and observe.
There should be no crash of LocalMediaPlayer.

Jira: None
Test: None