Fix issue that the position of QuickSetting is improper on DUT (IVI)
NotificationPanelView uses the width of notification stack Scroller to
estimate if notification stack Scroller and QuickSetting to be applied
the offset. That may cause the position of QuickSetting is improper if
the width of notification stack Scroller is smaller than the width of
QuickSetting.

The solution is that comparing the width of notification stack Scroller
with the width of QuickSetting and using the larger one to estimate if
notification stack Scroller and QuickSetting to be applied the offset.

Test:
1. Click on the top right of Status Bar.
2. Show notifications

Tracked-On: OAM-67772

Signed-off-by: Wang, ArvinX <arvinx.wang@intel.com>