mixin updates
egb_dv: Disable car diff config if disable_ipu is set to true
As ipu is not yet enabled in egb_dv, disable the car diff config
as it sets the ipu to static modules which causes device boot
issue.
egb_dv: Disable car diff config
As ipu is not yet enabled in egb_dv, disable the car diff config
as it sets the ipu to static modules which causes device boot
issue.

Change-Id: If1d1a19d3553269e05b721e928ae243f505f857c
Tracked-On: OAM-75540
Mixin-Reviewed-On: #/c/658853
Mixin-Reviewed-On: #/c/658854
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>