Adding only one pbap profile, client or server to the list
Issue : - We see that even when the pbab client profile
was present in the list, the pbab server profile was
added again in device details list. Therefore
contact sharing option for two pbap profiles server
and client appeared on UI.

Fix: - Keeping a check to not to add the pbap server
profile if pbab client was present in the device details
list.

Tracked-On: OAM-75968
Signed-off-by: Gaganpreet kaur <gaganpreetx.kaur@intel.com>