Fix data source for HEIF exif extractor
Do not allow the source to read past available bytes, since
the underlying input stream may not be able to seek after that.

For input streams from files, the available bytes is usually the
file size, and we usually don't have problem when the file is
of decent size. But when the file is very small, some of the
extractors (other than mp4) would request bytes past the end
of the file, which goes over the available range. Once that
condition is hit, we can't reset to the offet needed for
mp4 extractor and heif parsing would fail

Tracked-On: OAM-79903
Signed-off-by: varunjex <varunx.jeevangoudar@intel.com>
Signed-off-by: swaroopb <swaroop.balan@intel.com>