Fix race condition in btusb_sco_snd_card driver
Issue: Race condition occurs in the prepare call.
Prepare for playback and capture will call configure_endpoints
method to configure the TX and RX endpoints. When the prepare
is called for playback first, configure_endpoints method is
called which sets and BTUSB_ISOC_TX_EP_CONFIGURED and proceed
to do endpoint configuration. Prepare call can be called for
the capture during this time which calls the configure_endpoints
again. Since the BTUSB_ISOC_TX_EP_CONFIGURED is already set,
the configure_enpoints for the capture returns and it will
proceed to initialise the capture URBs. Since the endpoints
were in the process of configuration, the EP descriptors are
not getting populated properly at the time of initialising
capture URB, submission of capture URB to fail. Similarly
issue will also occur if capture is called first and prepare
is called next.

Fix: Added mutex locking for the configure_endpoints method

Tracked-On: OAM-81187
Signed-off-by: Aiswarya Cyriac <aiswarya.cyriac@intel.com>