[WA patch]Delay the system to go to sleep when screen off
Some CTS tests need turn off the screen via send KEY code SLEEP,
but because of the NUC don't support doze mode, receive SLEEP KEY
will cause the system to go to sleep immediately and make the test fail.
So, add a 30 seconds wakelock to delay the system to go to sleep when screen off
for these CTS tests.

Signed-off-by: Ting Li <ting.li@intel.com>