google_diff: fix the get ModemActivityTimeout when defaultPhone is null
Currently, the CTS case of BatteryStatsValidationTest#testRealtime is
failed
because it is always timeout in getting modem energy information in GP.

The root cause is no modem is present in GP then the information cannot
be getted
as the defaultphone is null.

to unblock beta miletone first. It will be submit to Google for a review
and merge.

Change-Id: I545b85ae10f83f0d43a73f8298a1bff2cbecaec2
Tracked-On: https://jira.devtools.intel.com/browse/OAM-84993
Signed-off-by: Zhen Han <zhen.han@intel.com>