Return the disabled state for selinux
Since SELinux is not enabled in cic now, there are many
APIs exsiting in frameworks need to take some actions basing
on the SELinux's state. They are often check the mount
point of /sys/fs/selinux, if it exists, then they think SELinux
is enabled. This is not a correct way for cic.

Since SELinux is not enabled in cic, so we always return the
disabled state for SELinux.

Tracked-On: OAM-89804
Signed-off-by: ji, zhenlong z <zhenlong.z.ji@intel.com>