Fix some CTS test case issues
The dumpable value of some processes may be affected by suid_dumpable
defined in /proc/sys/fs/suid_dumpable. For the processes that have extra
capabilities assigned through init.rc or fs config files, their dumpable
vaules are set to suid_dumpable.
The value of suid_dumpable of bare metal android is SUID_DUMP_DISABLE(0),
but this value maybe be changed to SUID_DUMP_ROOT(2) in ubuntu by some
host services, such as apport. This may introduce some regressions for the
processes of android in container, and it may also impact the CTS result.
So we need to set suid_dumpable's value back to SUID_DUMP_DISABLE if we detect
its value was set to SUID_DUMP_ROOT.

Tracked-On: OAM-90728
Signed-off-by: ji, zhenlong z <zhenlong.z.ji@intel.com>