mixin updates
Re-direct qemu_start_android log to a separate log
For automation, start_android_qcow2.sh is expected to run
qemu in the background and exit. Qemu logs prohibit the
start_android_qcow2.sh from exit after starting the
android in background.
so re-direct qemu_start_android log to a separate log when
--display-off/--use-adb and display log after android is
started in the background.

Tracked-On: OAM-90709
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/599
Signed-off-by: saddamhu-intel <saddam.hussain.abbas@intel.com>