Thermal: Add fixes for VTS
VTS expects temperature and threshold values in 2 ways.
If filterType argument is true, only temperature and
thresholds of that temperature type should be returned
by getCurrentTemperatures() and getTemperatureThresholds()
methods. CIV currently supports only CPU temperature
and emulated test battery temperature. Other temperatures
are not supported. This patch adds workaround to send
fake data for other temperature types to pass VTS.
Also, when filterType argument is false, VTS test expects
temperatures to be sent in order. So, CPU, GPU and battery
temperatures should be sent in order. GPU temperature
is missing in CIV. So, add a dummy entry for GPU.

Tracked-On: OAM-91198
Signed-off-by: Saranya Gopal <saranya.gopal@intel.com>