mixin updates
Enable selinux for android 11 on Celadon
Basing on the requirement, selinux should be enabled on
Celadon with android 11.
Change hwc & gralloc library name
This is to help change hwc & gralloc library name
from "intel" to "celadon". With this change, we
could delete some bsp diff
update drm interface version
update interface for clearkey, remove widevine which is not part
of caas by default.

Tracked-On: OAM-92961
Tracked-On: OAM-92984
Tracked-On: OAM-92991
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/886
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/894
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia/pull/1014
Signed-off-by: Ren Chenglei <chenglei.ren@intel.com>
Signed-off-by: Ruan, Hongfu <hongfu.ruan@intel.com>
Signed-off-by: ji, zhenlong z <zhenlong.z.ji@intel.com>