Fix signapk.jar not found in RELEASE_BUILD=true
When built with RELEASE_BUILD=true option builder tries to
call the signapk.jar file to generate the caas-sign***.zip file.
This signapk path is not set properly in Android 11.
The path is set relative to the script directory but it should
be relative to the output directory.
This patch fixes this issue and sets the path properly.

Tracked-On: OAM-92994
Signed-off-by: Tanuj Tekriwal <tanuj.tekriwal@intel.com>