disable fusectl for AIC
fusectl is a global fs for fuse, it can be used to control all the fuse
connections in the system, e.g. abort the fs, it is OK for the normal
system, but it will be a bad idea for the container use case, if one of
the container abort the fuse fs connection, all the other containers or
the host will be interrupted

the vold in Android R will abort all the fuse connections when stop, it
is horrile for AIC, the fuse fs created by host or aic-manager, e.g.
lxcfs, will be aborted, as well as other ones created by other AIC
instances, what a mess

that is the root cause of the lxcfs abort issue

so for AIC we should disable the fusectl

Tracked-On: ACP-1336
Topic: aic.bug.lxcfs-abort
Signed-off-by: Joe Zheng <joe.zheng@intel.com>
Change-Id: Ia8e0dd3c7845b8fc938f6f2009a12f49db89ae94