mixin updates
Switch allocator & mapper from 2.0 to 4.0
This to help enable allocator & mapper to 4.0, which is
required to support API 30.
Set the vintf manifest target-level to 5
This patch sets the vintf device manifest target-level=5
This is the minimum requirement for devices launching with
android 11. Where the product shipping api is set to >=30.

Tracked-On: OAM-95365
Tracked-On: OAM-95405
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1005
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1018
Signed-off-by: Ren Chenglei <chenglei.ren@intel.com>
Signed-off-by: Tanuj Tekriwal <tanuj.tekriwal@intel.com>