Change GL_CHECKPOINT level from LOW to MODERATE
When call swapBuffers in Skia, it fails to pass
GL_CHECKPOINT(LOW) check. DEBUG_OPENGL is defined
on Android R and set to DEBUG_LEVEL_LOW, then
DEBUG_OPENGL is equal to LOW, and return back with
"GL_INVALID". This seems not reasonable.

This change helps to fix CTS MultiDisplaySecurityTests
issue in CtsWindowManagerDeviceTestCases module

Tracked-On: OAM-95725
Signed-off-by: Ren Chenglei <chenglei.ren@intel.com>