mixin updates
set the property ro.vendor.hdmi.audio if it is TGL
get the cpu info and update audio property
ro.vendor.hdmi.audio
[WA]audio server restart on boot complete
Since hda driver is unable to get the HDMI
sound devices, it is waiting till the timeout
and then listing the ALC256 cards.
By then Android would have started the Audioserver
and finds that there are no sink devices availabe.
Till we get the HDMI solution working this can be
a WA fix for the audio 3.5mm playback

Tracked-On: OAM-95721
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1051
Mixin-Reviewed-On: https://github.com/projectceladon/device-androidia-mixins/pull/1056
Signed-off-by: gkdeepa g.k.deepa@intel.com