Update logic to program AUD_FREQ_CNTRL register based on new guidance.
Earlier this register was configured by BIOS and driver discovered the
value at init. This is no longer recommended and instead for driver
should set the values based on the hardware revision. This change
applies for all GEN12+ hardware.

Add the recommended values for all supported hardware. Extend the debug
print to also include values of the register as written by BIOS. This
can help debug rare cases where an older BIOS configures an incorrect
value.

Bspec: 49279

once the patch gets reviewed we need to back port the patch
http://intel-gfx-pw.fi.intel.com/series/7315/

Tracked-On: OAM-95960
Signed-off-by: Deepa G K <g.k.deepa@intel.com>