i915: Fix CFI violation in show_dynamic_id function
clang's Control Flow Integrity requires that every indirect call has a
valid target, which is based on the type of the function pointer. The
show_dynamic_id()functions is written as if it will be called
from dev_attr_show(); however, it will be called from
sysfs_kf_seq_show() because the files were created by
sysfs_create_group() and the sysfs ops are based on kobj_sysfs_ops
because of kobject_add_and_create(). Because the show_dynamic_id()
functions do not match the type of the show() member in struct
kobj_attribute, there is a CFI violation.

Tracked-On: OAM-100059
Signed-off-by: Patibandla, KiranX Kumar <kiranx.kumar.patibandla@intel.com>