WA Fix for BT suspend resume issue in Tasmania
In the specific hardware configuration power is cut to
the Bluetooth controller every time system enter S3 suspend

In some cases when frequent usb connect and disconnect
occurs, there is a chance that hci dev registration happens
even before cleanup is done for the existing hci device. Delay
is added in this case to allow enough time for cleanup of existing
device before registering a new one.

Tracked-On: OAM-99203
Signed-off-by: Aiswarya Cyriac <aiswarya.cyriac@intel.com>