Enable virtual ab compression ota
Need this patch to enable virtual ab compression feature,
or the system cannot boot up after the OTA completes.

Simply speaking, snapuserd is needed to be started before
the init second stage, since system partition need to be
setup after the snapuserd is started, or the system partition
cannot be accessed. However, init second stage needs to
access the system partition, So add this patch to make sure
snapuserd is started before init second stage.

Will report a bug to Google, and this patch also will be removed
once this issue is resovled.

Tracked-On: OAM-100724
Signed-off-by: ji, zhenlong z <zhenlong.z.ji@intel.com>