Fix for Camera Asking Location Permission every time
Camera app prompts for "Tag your photos and videos with the locations
where they're taken" permission on every open.

If the lastversion value is less than 2, it will force the location
choice to user on every camera open.

Set the lastversion value to 2 to skip forcing the location choice
to user.

Change-Id: Icff3ca5db8a2c34732885694d620f51dd22d2931
Tracked-On: OAM-104683
Signed-off-by: Singh, Sapna1 <sapna1.singh@intel.com>