egl: return if context not changed for eglMakeCurrent
Issue:
This issue was exposed during test when VDS was killed suddenly,
eglMakeCurrent failed and frameworks would try to fallback to the
previous context, which is still the current context, and re-set the
same context cause a dead lock in dequeueBuffer.

Cause:
Sometimes set the same context again would cause a dead lock in
dequeueBuffer, since the buffer already acquired by it's consumer
and not released yet.

Solution:
Do nothing in eglMakeCurrent if context and surface not changed
to avoid this kind of issue.

Change-Id: I92a25a6cdbdda2a662c3abc326807d30adbe9149
Tracked-On: VVP0-3694
Signed-off-by: Shuo Gao <shuo.gao@intel.com>