Increase number of patches limit from 100 to 500 for autopatch
As of now autopatch only checks for 100 patches in a folder to apply
"Already applied" tag. Increasing the same to 500 as the number of
Patches in framework-base folder is more than 100. This will enable
Lunch to not try and apply already applied patches till the number of
Patches in a folder reaches 500+.

Tracked-On: OAM-105792
Signed-off-by: Reddy, Alavala Srinivasa <alavala.srinivasa.reddy@intel.com>