Fix delayed loading and distorted thumbnails observed in Gallery App
With few files, thumbnails are distorted.

Software codec is used for thumbnails. Using Codec2.0 software
codecs is resulting in delay in loading the thumbnail and also
some of the thumbnails are distorted.

Fix the issue by enabling Codec2.0 decoder only for AV1. For
rest of the formats, OMX decoder is used.

Tracked-On: OAM-105931
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>