Resolution set in Camera lost[back operation]
Changes done in camera app was to check the working of end to end
dynamic client update solution.

User seleted resoultion setting is not preserved if we keep the
changes in camera app for dynamic client update solution. So,
reverting the app changes so resolution setting is preserved and
behavior is same when compared with other camera apps.

Tracked-On: OAM-106572
Signed-off-by: Singh, Sapna1 <sapna1.singh@intel.com>