Update bthfp call state to audio manager
As we are using btusb loopback device, there is no
audio output if we open btusb card during DAILING/RINGING
call state. Audio is routed properly when btusb card is
opened after call is picked up (i.e., call state to ACTIVE).

We are updating the same to audio manager.