no need to exec the wait_for_keymaster
This process may increase the boot time due to it will
check the keystore service status before doing mount,
however, in the latest code it won't need anymore because
the init will listen for property changes on a separate thread.
Hence, removing this process can help reduce the boot time.

Tracked-On: OAM-111202
Signed-off-by: Yan, Shaopu <shaopu.yan@intel.com>