Removing fixes done for CVE-2023-21111 & CVE-2023-21267
CVE-2023-21111 - The May 2023 Android Automotive OS Security Bulletin
Partner Preview has not been changed.
So, it requires additional time for testing and verification

CVE-2023-21267 was added for 2023-08-01 due to a potential functional
regression that is known to impact screen unlocking by causing any
password entered in the pinned app interface to unlock the screen.
ASB recommend Partners remove this patch.

Tracked-On: OAM-112236
Signed-off-by: Alam, SahibeX <sahibex.alam@intel.com>