REVERTME: Increase number of patches limit for apply_patch
As of now apply_patch only checks for 200 patches in a folder to apply
"Already applied" tag. Increasing the same to 250 as the number of
Patches in framework-base folder is more than 200. This will enable
Lunch to not try and apply already applied patches till the number of
Patches in a folder reaches 250+.

Tracked-On: OAM-112678
Signed-off-by: Reddy, Alavala Srinivasa <alavala.srinivasa.reddy@intel.com>