ftrace: Ignore f2fs/f2fs_truncate_partial_nodes.nid field
The field has a different format on kernels before and after
0b04d4c0542e("f2fs: Fix f2fs_truncate_partial_nodes ftrace event").
Perfetto is not able to understand the new format anyway and it would
hit a DCHECK (abort in non debug builds).

Skipped test code for it's too new to merge into current perfetto(v15.0).

Apply to aosp_diff/preliminary folder.

Tests done: perfetto traced will not crash with this patch

Tracked-On: OAM-113067
Signed-off-by: chenyanxzhu <chenyanx.zhu@intel.com>