Revert aosp l2cap/smp patches to fix BT connection issue
[Problem]
In SMP layer, the acknowledgement for packets being sent is not checked
properly, causing the SMP control block to reset, which is flushing the
unsent smp packets in l2cap queue. This packet drop is causing the remote
reference device to disconnect with DUT as pairing is incomplete.
SMP signing info cmd is flushed in l2cap layer.

[Solution]
So SMP to register the tx_complete_cb with l2cap for getting the
acknowledgement for sent pkts. Wait for all sent pkts acknowledgement,
until then do not reset the SMP control block.
Now SMP signing info cmd is being sent to reference device and able to
see DUT in reference device paired devices list.

Tracked-On: OAM-112831
Signed-off-by: Gowtham Anandha Babu <gowtham.anandha.babu@intel.com>
Signed-off-by: Bharat B Panda <bharat.b.panda@intel.com>