Enable AX210 and AX201 Bluetooth for Android Automotive
The BT chip is loaded as USB interface as hci0, the UsbHostManager
is taking control of hci0 interface and doing initialization.
Because of which Bluetooth-HAL is not able to send HCI cmds via
hci0 interface, causing the hci_timeout_abort() crash.
So BT is not enabled.

Add BT chip vendor id and product id in config deny list, so that
UsbHostManager skips the USB interface intialization for hci0.

Tests:
Open bluetooth in Settings app,
bluetooth device can be found.

Tracked-On: OAM-118006
Signed-off-by: Gowtham Anandha Babu <gowtham.anandha.babu@intel.com>
Signed-off-by: Bharat B Panda <bharat.b.panda@intel.com>