Fix the error of sound output disorder
System will reuse the ringtone on multi-passenger mode when the ringtone
has been used, but the sould will be closed when finishing to play,
the sould will be played on remote and on another zone. so we always
stop old sound and start to play new sound when adjusting the sound
volumn.

Tracked-On: OAM-133048
Signed-off-by: Xu Bing <bing.xu@intel.com>