Fix the wrong routing when playing feedback ringtone
Create the MediaPlayer instance for the Ringtone object before
calling play() to avoid incorrect logic being triggered during
ringtone playback. If the Ringtone is not created at this point,
calling play() will check whether to use a RemotePlayer. The
RemotePlayer is intended for certain system apps that want to play
ringtones within the SystemUI process by invoking the setRemote()
interface, which delegates playback to the system server side.

Tracked-On: OAM-133236
Signed-off-by: Wan, Xinxin <xinxin.wan@intel.com>
Signed-off-by: Xu Bing <bing.xu@intel.com>